Добавил TestValues widget
