+++ lib/src/features/settings/widget/screens/settings_screen.dart
@@ -1,15 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:insight/gen/pubspec.yaml.g.dart';
 import 'package:insight/src/common/constants/app_strings.dart';
 import 'package:insight/src/common/constants/base_constants.dart';
-import 'package:insight/src/common/utils/build_mode.dart';
 import 'package:insight/src/common/utils/current_flavor.dart';
-import 'package:insight/src/common/utils/extensions/context_extension.dart';
 import 'package:insight/src/common/utils/extensions/go_relative_named.dart';
 import 'package:insight/src/common/widgets/adaptive_scaffold.dart';
 import 'package:insight/src/common/widgets/buttons/adaptive_button.dart';
 import 'package:insight/src/common/widgets/app_bars/custom_app_bar.dart';
+import 'package:insight/src/common/widgets/test_values_widget.dart';
 import 'package:insight/src/features/auth/bloc/auth_bloc.dart';
 import 'package:insight/src/features/profile/bloc/profile_bloc.dart';
 import 'package:insight/src/features/settings/widget/components/profile_widget.dart';
@@ -81,7 +79,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
                   onTap: () => context.goRelativeNamed('about'),
                 ),
                 const SizedBox(height: 20),
-                if (!Flavor.isProd) const _TestValues(),
+                if (!Flavor.isProd) const TestValues(),
                 const SizedBox(height: 20),
                 if (isAuthenticated)
                   AdaptiveButton(
@@ -97,61 +95,6 @@ class _SettingsScreenState extends State<SettingsScreen> {
             );
           },
         ),
-      ),
-    );
-  }
-}
-
-class _TestValues extends StatelessWidget {
-  const _TestValues();
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.symmetric(vertical: 16),
-      decoration: ShapeDecoration(
-        color: context.colorScheme.surfaceContainerLowest,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(20),
-        ),
-      ),
-      child: Column(
-        children: [
-          _TestValueText(title: 'flavor', value: Flavor.current),
-          _TestValueText(title: 'build mode', value: BuildMode.current),
-          _TestValueText(
-            title: 'version',
-            value: Pubspec.version.representation,
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _TestValueText extends StatelessWidget {
-  const _TestValueText({
-    required this.title,
-    required this.value,
-  });
-
-  final String title;
-  final String value;
-
-  @override
-  Widget build(BuildContext context) {
-    return RichText(
-      text: TextSpan(
-        text: '$title: ',
-        children: <TextSpan>[
-          TextSpan(
-            text: value,
-            style: DefaultTextStyle.of(context).style.copyWith(
-                  color: context.colorScheme.primary,
-                  fontWeight: FontWeight.bold,
-                ),
-          ),
-        ],
       ),
     );
   }
